+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added Jeep 4.0L sensor family (gauge temp sender, Renix ECU CTS, oil pressure sender)
+
 ## [0.7.0-beta] - 2026-02-07
 
 ### Added